Delete CardWrapper shared components - Move components to correct places
