+++ webapp/src/components/Card/Card.js
@@ -1,9 +1,13 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import { CardWrapper, CardActionsWrapper, CardActionAreaWrapper, CardContentWrapper } from '../../shared/CardWrapper';
 import CardMedia from '@material-ui/core/CardMedia';
 import Typography from '@material-ui/core/Typography';
 import { StyledBox } from '../../components';
+import styled from 'styled-components';
+import Card from '@material-ui/core/Card';
+import { spacing, palette, typography } from '@material-ui/system';
+import CardActionArea from '@material-ui/core/CardActionArea';
+import CardContent from '@material-ui/core/CardContent';
 
 const useStyles = makeStyles({
 	root: {
@@ -14,15 +18,16 @@ const useStyles = makeStyles({
 		height: 150
 	}
 });
-const CardElement = (props) => {
+
+const MunchiesCard = (props) => {
 	let priceRange = '';
 	for (let i = 0; i < props.price_range; i++) {
 		priceRange += '$';
 	}
 	const classes = useStyles();
 	return (
-		<CardWrapper color={'primary'} className={classes.root}>
-			<CardActionAreaWrapper>
+		<Wrapper color={'primary'} className={classes.root}>
+			<CardActionAreas>
 				<CardMedia className={classes.media} image={props.image} title={props.name} />
 				<CardContentWrapper bgcolor={'cardBackgroundColor'}>
 					<Typography variant="body2" color={'primary'} component="div">
@@ -35,9 +40,31 @@ const CardElement = (props) => {
 						</StyledBox>
 					</Typography>
 				</CardContentWrapper>
-			</CardActionAreaWrapper>
-		</CardWrapper>
+			</CardActionAreas>
+		</Wrapper>
 	);
 };
 
-export default CardElement;
+const Wrapper = styled(Card)`
+  ${spacing};
+  ${palette};
+`;
+
+const CardActionAreas = styled(CardActionArea)`
+  ${spacing};
+  ${palette};
+  ${typography};
+
+  &.MuiCardActionArea-root {
+    /* display: flex; */
+    justify-content: center;
+  }
+`;
+
+const CardContentWrapper = styled(CardContent)`
+  ${spacing};
+  ${palette};
+  ${typography};
+`;
+
+export default MunchiesCard;
